Used various tricks to improve generic runtime performance.

Results on my machine as follows:

- before the change

```
decode vector tile with pbf x 858 ops/sec ±1.17% (90 runs sampled)
encode vector tile with pbf x 556 ops/sec ±1.46% (84 runs sampled)
```

- after the change

```
decode vector tile with pbf x 999 ops/sec ±1.13% (92 runs sampled)
encode vector tile with pbf x 638 ops/sec ±1.48% (84 runs sampled)
```

(both show ~15% improvement)