@mourner 2016-05-26
Brilliant! Will be happy to merge this one.